fix drdy event clearing
In AH_BIO case the drdy event is cleared when the high
byte of at least one of the axis is read, so the procedure
reads and discard also Z_H byte. This is not required for
vafe-only case. The status of this case is kept in the
ctx private structure to avoid an extra byte bus reading.

Signed-off-by: Armando Visconti <armando.visconti@st.com>